src/backend/requirements/common_frozen.txt
annotated-types==0.6.0
asgiref==3.8.1
asttokens==2.4.1
bcrypt==4.1.2
black==24.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
decorator==5.1.1
Deprecated==1.2.14
Django==5.0.3
django-debug-toolbar==4.3.0
django-dramatiq==0.11.6
django-environ==0.11.2
django-extensions==3.2.3
django-ninja==1.1.0
django-pdb==0.6.2
django_periodiq @ git+https://github.com/Sovetnikov/django_periodiq@3d6f18fe8339badb83716161476559e27bf16882
dramatiq==1.16.0
executing==2.0.1
fabric==3.2.2
Faker==24.4.0
gevent==24.2.1
greenlet==3.0.3
idna==3.6
invoke==2.2.0
ipdb==0.13.13
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==24.0
paramiko==3.4.0
parso==0.8.3
pathspec==0.12.1
pendulum==2.1.2
periodiq==0.12.1
pexpect==4.9.0
pika==1.3.2
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytzdata==2020.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.10.0
urllib3==2.2.1
watchdog==4.0.0
watchdog-gevent==0.1.1
wcwidth==0.2.13
wrapt==1.16.0
zope.event==5.0
zope.interface==6.2


